major refactor
fixes #58 and #57
