+++ fmi-base/src/main/kotlin/no/ntnu/ihb/fmi4j/modeldescription/jacskon/JacksonSourceFile.kt
@@ -1,0 +1,14 @@
+package no.ntnu.ihb.fmi4j.modeldescription.jacskon
+
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
+import no.ntnu.ihb.fmi4j.modeldescription.SourceFile
+
+/**
+ * @author Lars Ivar Hatledal
+ */
+data class JacksonSourceFile(
+
+        @JacksonXmlProperty
+        override val name: String
+
+) : SourceFile
