+++ fmi-base/src/main/kotlin/no/ntnu/ihb/fmi4j/modeldescription/variables/ModelVariables.kt
@@ -24,13 +24,10 @@
 
 package no.ntnu.ihb.fmi4j.modeldescription.variables
 
-import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper
-import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
-import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement
+
 import no.ntnu.ihb.fmi4j.common.StringArray
 import no.ntnu.ihb.fmi4j.common.ValueReference
 import no.ntnu.ihb.fmi4j.common.ValueReferences
-import java.io.Serializable
 
 /**
  * @author Lars Ivar Hatledal
@@ -117,32 +114,6 @@ interface ModelVariables : Iterable<TypedScalarVariable<*>> {
      */
     fun getByCausality(causality: Causality): List<TypedScalarVariable<*>> {
         return filter { it.causality == causality }
-    }
-
-}
-
-/**
- * @author Lars Ivar Hatledal
- */
-@JacksonXmlRootElement(localName = "ModelVariables")
-class ModelVariablesImpl : ModelVariables, Serializable {
-
-    @JacksonXmlProperty(localName = "ScalarVariable")
-    @JacksonXmlElementWrapper(useWrapping = false)
-    private val variables: List<ScalarVariableImpl>? = null
-
-    @Transient
-    private var _variables: List<TypedScalarVariable<*>>? = null
-
-    override fun getVariables(): List<TypedScalarVariable<*>> {
-        if (_variables == null) {
-            _variables = variables!!.map { it.toTyped() }
-        }
-        return _variables!!
-    }
-
-    override fun toString(): String {
-        return "ModelVariablesImpl(variables=$variables)"
     }
 
 }
