+++ fmi-base/src/test/kotlin/no/ntnu/ihb/fmi4j/modeldescription/vendors/openmodelica/FmuExportCrossCompileTest.kt
@@ -1,16 +1,11 @@
 package no.ntnu.ihb.fmi4j.modeldescription.vendors.openmodelica
 
 import no.ntnu.ihb.fmi4j.modeldescription.TestFMUs
-import no.ntnu.ihb.fmi4j.modeldescription.ModelDescription
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import java.io.ByteArrayInputStream
-import java.io.ByteArrayOutputStream
-import java.io.ObjectInputStream
-import java.io.ObjectOutputStream
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class FmuExportCrossCompileTest {
@@ -34,12 +29,12 @@ class FmuExportCrossCompileTest {
         val d1 = der[0]
         Assertions.assertEquals(d1.index, 3)
         Assertions.assertEquals(d1.dependencies.size, 1)
-        Assertions.assertEquals("dependent", d1.dependenciesKind)
+        Assertions.assertEquals(listOf("dependent"), d1.dependenciesKind)
 
         val d2 = der[1]
         Assertions.assertEquals(d2.index, 4)
         Assertions.assertTrue(d2.dependencies.isEmpty())
-        Assertions.assertEquals("", d2.dependenciesKind)
+        Assertions.assertEquals(emptyList<String>(), d2.dependenciesKind)
     }
 
     @Test
@@ -47,24 +42,6 @@ class FmuExportCrossCompileTest {
         val variableNamingConvention = modelDescription.variableNamingConvention
         LOG.info("variableNamingConvention=$variableNamingConvention")
         Assertions.assertTrue(modelDescription.variableNamingConvention == "structured")
-    }
-
-    @Test
-    fun testSerialization() {
-
-        val bos = ByteArrayOutputStream()
-        ObjectOutputStream(bos).use {
-            it.writeObject(modelDescription)
-            it.flush()
-        }
-
-        ObjectInputStream(ByteArrayInputStream(bos.toByteArray())).use {
-            val md: ModelDescription = it.readObject() as ModelDescription
-            Assertions.assertEquals(modelDescription.modelVariables.size, md.modelVariables.size)
-            md.modelVariables.forEach { LOG.info("$it") }
-            LOG.info("${md.modelStructure}")
-        }
-
     }
 
 }
