+++ fmi-base/src/test/kotlin/no/ntnu/ihb/fmi4j/modeldescription/vendors/twentysim/ControlledTemperatureTest.kt
@@ -24,8 +24,8 @@
 
 package no.ntnu.ihb.fmi4j.modeldescription.vendors.twentysim
 
+import no.ntnu.ihb.fmi4j.modeldescription.DefaultExperiment
 import no.ntnu.ihb.fmi4j.modeldescription.TestFMUs
-import no.ntnu.ihb.fmi4j.modeldescription.misc.DefaultExperiment
 import no.ntnu.ihb.fmi4j.modeldescription.variables.ModelVariables
 import no.ntnu.ihb.fmi4j.modeldescription.variables.RealVariable
 import org.junit.jupiter.api.Assertions
