+++ fmu2jar/src/test/kotlin/no/ntnu/ihb/fmu2jar/codegen/CodeGenerationTest.kt
@@ -1,9 +1,8 @@
 package no.ntnu.ihb.fmu2jar.codegen
 
-import no.ntnu.ihb.fmi4j.modeldescription.parser.ModelDescriptionParser
+import no.ntnu.ihb.fmi4j.modeldescription.jacskon.JacksonModelDescriptionParser
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.File
@@ -20,7 +19,7 @@ class CodeGenerationTest {
         val file = File("../fmus/2.0/cs/20sim/4.6.4.8004/ControlledTemperature/ControlledTemperature.fmu")
         Assertions.assertTrue(file.exists())
 
-        ModelDescriptionParser.parse(file).also {
+        JacksonModelDescriptionParser.parse(file).also {
             LOG.info( CodeGenerator(it).generateBody() )
         }
